+++ src/ui_ng/lib/src/create-edit-endpoint/create-edit-endpoint.component.spec.ts
@@ -21,6 +21,7 @@ describe('CreateEditEndpointComponent (inline template)', () => {
     "name": "target_01",
     "username": "admin",
     "password": "",
+    "insecure": false,
     "type": 0
   };