+++ tools/migration/changelog.md
@@ -55,3 +55,4 @@ Changelog for harbor database schema
   - create table `project_metadata`
   - insert data into table `project_metadata`
   - delete column `public` from table `project`
+  - add column `insecure` to table `replication_target`